blog-repo/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
distlib==0.3.8
Django==5.1.1
filelock==3.16.0
h5py==3.11.0
markdown-it-py==3.0.0
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==2.1.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
platformdirs==4.3.2
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.7.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.4
